+++ .gitignore
@@ -1,3 +1,2 @@
 *.pyc
 __pycache__/
-data
