--- tuepa/data/__init__.py
@@ -1,1 +1,0 @@
-from .preprocessing import read_passages
