--- tuepa/data/elmo/__init__.py
@@ -1,2 +1,0 @@
-from .elmo_input import get_elmo_input_fn
-from .elmo_processing import preprocess_dataset, specific_elmo
