--- tuepa/parser/__init__.py
@@ -1,2 +1,0 @@
-from .states.state import State
-from .oracle import Oracle
