--- tuepa/util/__init__.py
@@ -1,2 +1,0 @@
-from .util import SaverHook
-from .numberer import *
